Gold/requirements.txt
# AI Employee Project — Gold Tier Dependencies

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
watchdog>=3.0.0

# Gmail API
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0

# LinkedIn (browser automation)
playwright>=1.40.0

# Scheduling
schedule>=1.2.0

# MCP Server
# (uses stdlib json + sys for JSON-RPC over stdio)

# Email sending
# (uses stdlib smtplib + email)

# Odoo ERP
odoorpc>=0.10.1

# Social Media APIs
tweepy>=4.14.0
instagrapi>=2.1.0
facebook-sdk>=3.1.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
colorama>=0.4.6
